requirements.txt
anthropic>=0.39.0
websocket-client>=1.6.0
asyncio>=3.4.3
aiohttp>=3.9.0
click>=8.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
colorama>=0.4.6
requests>=2.31.0
certifi>=2023.0.0
urllib3>=2.0.0